Updated PSP Panel
took loc list out of chat panel and added it to psp panel.
