+++ src/Tests/LOCTest.java
@@ -1,0 +1,21 @@
+package Tests;
+
+import static org.junit.Assert.*;
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+import net.sf.memoranda.LinesofCode;
+import net.sf.memoranda.LOCList;
+import net.sf.memoranda.LOCListImpl;
+
+public class LOCTest {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
