+++ src/net/sf/memoranda/ui/PSPPanel.java
@@ -7,6 +7,8 @@ import java.awt.Dimension;
 import java.awt.Font;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.util.ArrayList;
+import java.util.Vector;
 
 import javax.swing.DefaultCellEditor;
 import javax.swing.JButton;
@@ -15,12 +17,17 @@ import javax.swing.JComboBox;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
+import javax.swing.JTextField;
 import javax.swing.JToolBar;
+import javax.swing.table.AbstractTableModel;
 import javax.swing.table.TableColumn;
 import javax.swing.table.TableModel;
 
 import net.sf.memoranda.CurrentProject;
 import net.sf.memoranda.DefectList;
+import net.sf.memoranda.LOCList;
+import net.sf.memoranda.LinesofCode;
+import net.sf.memoranda.LOCList;
 import net.sf.memoranda.History;
 import net.sf.memoranda.LOCList;
 import net.sf.memoranda.NoteList;
@@ -29,6 +36,7 @@ import net.sf.memoranda.Project;
 import net.sf.memoranda.ProjectListener;
 import net.sf.memoranda.ResourcesList;
 import net.sf.memoranda.TaskList;
+import net.sf.memoranda.date.CalendarDate;
 import net.sf.memoranda.ui.DefectPanel.DefectsTableModel;
 import net.sf.memoranda.util.Local;
 
@@ -41,10 +49,18 @@ public class PSPPanel extends JPanel {
 	JButton historyBackB = new JButton();
 	JButton historyForwardB = new JButton();
 	//Defect Log Buttons
-	JButton addRowBtn = new JButton("Add Row");
-	JButton deleteBtn = new JButton("Delete");
+	JButton addRowBtn = new JButton("Add Defect");
+	JButton deleteBtn = new JButton("Delete Defect");
 
 	JScrollPane scrollPane = new JScrollPane();
+	
+	// Lines of Code Panel stuff
+			LinesofCode loc_object = null;
+			LOCList loc_log = null;
+			JTextField locField = new JTextField(8);
+			JButton addLOC = new JButton("Add Lines of Code");
+			JPanel locPanel = new JPanel();
+			JButton deleteLOC =  new JButton("Delete loc");
 
 	DailyItemsPanel parentPanel = null;
 
@@ -105,7 +121,42 @@ public class PSPPanel extends JPanel {
 
 		toolBar.add(addRowBtn);
 		toolBar.add(deleteBtn);
+		toolBar.add(addLOC);
+		toolBar.add(deleteLOC);
+		
 		this.add(toolBar, BorderLayout.NORTH);	
+		
+		
+		LocTableModel tableModelloc = new LocTableModel();
+		JTable tableloc = new JTable(tableModelloc);
+		JScrollPane scrollPaneloc = new JScrollPane(tableloc);
+		
+		tableloc.getColumnModel().getColumn(0).setPreferredWidth(20*1000);		
+		tableloc.getColumnModel().getColumn(1).setPreferredWidth(50*1000);
+		tableloc.getColumnModel().getColumn(2).setPreferredWidth(100*1000);
+		tableloc.getColumnModel().getColumn(3).setPreferredWidth(200*1000);
+		//TableColumn col = table.getColumnModel().getColumn();
+		
+		this.add(scrollPaneloc, BorderLayout.SOUTH);
+
+		tableloc.setFillsViewportHeight(true);
+		tableModelloc.populateTable();
+
+		// The Add LOC button adds a new row to the table
+		addLOC.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent arg0) {
+				tableModelloc.addNewRow();
+			}
+		});
+		
+		deleteLOC.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent arg0) {
+				tableModelloc.deleteRows(tableloc.getSelectedRows());
+			}
+		});
+		
+		
+			
 		
 		//Defect Log Table Creator	
 		DefectPanel defectLog = new DefectPanel();
@@ -187,6 +238,7 @@ public class PSPPanel extends JPanel {
 			}
 			public void projectWasChanged() {
 				((DefectsTableModel) tableModel).populateTable();
+				tableModelloc.populateTable();
 			}
 		}); 
 
@@ -202,7 +254,7 @@ public class PSPPanel extends JPanel {
 		timerPanel.add(pause);
 		timerPanel.add(reset);
 
-		this.add(scrollPane3, BorderLayout.SOUTH); //add scrollPane4 first when complete
+		this.add(scrollPane3, BorderLayout.EAST); //add scrollPane4 first when complete
 	}
 	
 	public class starts implements ActionListener{
@@ -268,5 +320,138 @@ public class PSPPanel extends JPanel {
 			  time.setText(String.valueOf(String.valueOf((dTime/1000)/60) + ":" 
 					  + String.valueOf((dTime/1000)%60) + ":" + String.valueOf((dTime)%1000)));
 		  } 
+	}
+	public class LocTableModel extends AbstractTableModel {
+		/**
+		 * 
+		 */
+		private static final long serialVersionUID = 1L;
+
+		// Table data
+		private ArrayList<Object[]> _data;
+
+		// Next loc 
+		private int _nextNumber;  
+
+		// Column names
+		private String[] ColumnNames = {
+				"Number",
+				"Date Completed",
+				"Lines of Code",
+				"Description"
+		};
+
+		// Column types
+		private Class[] ColumnClasses = {
+				Integer.class,
+				CalendarDate.class,
+				Integer.class,
+				String.class
+		};
+
+		public LocTableModel() {
+			_data = new ArrayList<Object[]>();
+			_nextNumber = 1; // needs to be loc
+		}
+
+		@Override
+		public int getRowCount() {
+			return _data.size();
+		}
+
+		@Override
+		public int getColumnCount() {
+			return ColumnNames.length;
+		}
+
+		@Override
+		public Object getValueAt(int row, int column) {
+			return _data.get(row)[column];
+		}
+
+		@Override
+		public Class getColumnClass(int c) {
+			return ColumnClasses[c];
+		}
+
+		@Override
+		public String getColumnName(int c) {
+			return ColumnNames[c];
+		}
+
+		@Override
+		public boolean isCellEditable(int row, int col) {
+			return (col != 0 && col != 1);
+		}
+
+		@Override
+		public void setValueAt(Object val, int row, int col) {
+			_data.get(row)[col] = val;
+
+			String attrib = "";
+			String value = "";
+			if (col == 2){
+				attrib = "lines_of_code";
+				value = Integer.toString((Integer)val);
+			}
+			if (col == 3)
+			{
+				attrib = "description";
+				value = (String)val;
+			}
+
+			CurrentProject.getLOCList().modifyLoc((Integer)_data.get(row)[0], attrib, value);;
+			fireTableDataChanged();
+		}
+		
+		/**
+		 * Add a new row for a new loc: has default data inputs.
+		 */
+		public void addNewRow() {
+			_data.add(new Object[] {
+					_nextNumber, 
+					new CalendarDate(),
+					0,
+					""
+			});
+
+			LinesofCode temp_loc = new LinesofCode();
+			temp_loc.setDateCompleted(CalendarDate.today());
+			temp_loc.setLocNumber(_nextNumber++); 
+			CurrentProject.getLOCList().addLOC(temp_loc);
+
+			fireTableDataChanged();
+		}
+
+		
+		public void deleteRows(int[] rowIndexes) {
+			System.out.println(rowIndexes.length);
+			for(int i = rowIndexes.length - 1; i >= 0; i--) {
+				CurrentProject.getLOCList().removeLOC((Integer)_data.get(rowIndexes[i])[0]);
+				_data.remove(rowIndexes[i]);
+			}
+
+			fireTableDataChanged();
+		}
+		
+		protected void populateTable() {
+			_data.clear();
+			Vector<LinesofCode> loc = CurrentProject.getLOCList().getAllLOC();
+			for (int i = 0; i < loc.size(); ++i)
+			{
+
+				_data.add(new Object[] {
+						loc.get(i).getLocNumber(),
+						loc.get(i).getDateCompleted(),
+						loc.get(i).getLOCIntValue(),
+						loc.get(i).getDescription()
+				});
+				if (i == loc.size() - 1)
+				{
+					_nextNumber = loc.get(i).getLocNumber() + 1;
+				}
+			}
+			fireTableDataChanged();
+		}
 	}
 }
